Move userCanModifyDandiset to landing page
